test: replace FSST overflow test with direct build_views rollover test
Drop the FSST-based fsst_canonicalize_offsets_overflow_i32 test in favor of
a direct build_views test in vortex-array, avoiding the compressor/training
and the doubled input+decompressed heaps. build_views_offsets_overflow_i32
builds a ~2.25 GiB heap (just past i32::MAX) and calls build_views with
MAX_BUFFER_LEN, asserting the heap rolls over into multiple buffers, that no
buffer exceeds MAX_BUFFER_LEN, and that rows straddling the rollover
boundary reconstruct exactly.

Validated that the test catches the overflow: with the single-buffer
fast-path guard intact it passes, and forcing the fast path to swallow the
whole heap makes it fail on the multi-buffer assertion.

Gated with #[test_with::env(CI)] / #[test_with::no_env(VORTEX_SKIP_SLOW_TESTS)],
matching the existing slow-test convention; adds test-with as a vortex-array
dev-dependency.
