+++ encodings/fsst/src/tests.rs
@@ -10,7 +10,6 @@ use vortex_array::LEGACY_SESSION;
 use vortex_array::VortexSessionExecute;
 use vortex_array::arrays::VarBinViewArray;
 use vortex_array::arrays::varbin::builder::VarBinBuilder;
-use vortex_array::arrays::varbinview::build_views::MAX_BUFFER_LEN;
 use vortex_array::assert_arrays_eq;
 use vortex_array::assert_nth_scalar;
 use vortex_array::dtype::DType;
@@ -168,93 +167,4 @@ fn fsst_compress_offsets_overflow_i32() {
     println!("compressing to FSST");
     let compressed = fsst_compress(array, len, &dtype, &compressor, &mut ctx);
     assert_eq!(compressed.len(), len);
-}
-
-// TODO(someone): ideally CI would run this in release mode as well since debug builds make the
-// allocation and decompression loop substantially slower.
-/// Decode-side companion to [`fsst_compress_offsets_overflow_i32`]. Canonicalizing an FSST array
-/// decompresses the whole heap and feeds it to `build_views` with `max_buffer_len = MAX_BUFFER_LEN`
-/// (`i32::MAX`). `build_views` has a single-buffer fast path taken only when the decoded heap fits
-/// within one buffer; when the heap exceeds `MAX_BUFFER_LEN` it must instead roll over into multiple
-/// buffers (resetting the per-buffer offset) so the `u32` view offsets never overflow.
-///
-/// This builds an FSST array whose *decompressed* size crosses `i32::MAX`, so canonicalization is
-/// forced down the rollover path. We assert the row count is preserved, that more than one data
-/// buffer was produced (proving the fast path was declined), that no single buffer exceeds
-/// `MAX_BUFFER_LEN`, and that values on both sides of the rollover boundary reconstruct exactly.
-///
-/// Highly compressible bodies keep the FSST array tiny, so peak memory is dominated by the ~2.25 GiB
-/// input and the ~2.25 GiB decompressed heap. Gated to CI and skipped when `VORTEX_SKIP_SLOW_TESTS`
-/// is set. To run it locally:
-///
-/// ```text
-/// CI=1 cargo test --release -p vortex-fsst fsst_canonicalize_offsets
-/// ```
-#[test_with::env(CI)]
-#[test_with::no_env(VORTEX_SKIP_SLOW_TESTS)]
-fn fsst_canonicalize_offsets_overflow_i32() {
-    const STRING_LEN: usize = 64 * 1024;
-    // Comfortably past MAX_BUFFER_LEN (`i32::MAX` ~= 2.0 GiB) so the decoded heap must roll over.
-    const TOTAL_BYTES: usize = (1usize << 31) + (256 << 20); // ~2.25 GiB
-    const N: usize = TOTAL_BYTES / STRING_LEN;
-
-    // Each value is a long, trivially compressible body carrying its row index as an ASCII prefix,
-    // so FSST compresses the heap to almost nothing while every row stays individually verifiable.
-    fn nth_string(i: usize) -> Vec<u8> {
-        let mut s = vec![b'x'; STRING_LEN];
-        let prefix = format!("row-{i:08}-");
-        s[..prefix.len()].copy_from_slice(prefix.as_bytes());
-        s
-    }
-
-    println!("building large VarBinArray");
-    let mut builder = VarBinBuilder::<i64>::with_capacity(N);
-    let mut buf = vec![b'x'; STRING_LEN];
-    for i in 0..N {
-        let prefix = format!("row-{i:08}-");
-        buf[..prefix.len()].copy_from_slice(prefix.as_bytes());
-        builder.append_value(&buf);
-    }
-    let array = builder.finish(DType::Utf8(Nullability::NonNullable));
-
-    println!("training FSST compressor");
-    let compressor = fsst_train_compressor(&array);
-    let len = array.len();
-    let dtype = array.dtype().clone();
-    let mut ctx = LEGACY_SESSION.create_execution_ctx();
-
-    println!("compressing to FSST");
-    let compressed = fsst_compress(array, len, &dtype, &compressor, &mut ctx).into_array();
-
-    println!("canonicalizing to VarBinView");
-    let canonical = compressed.execute::<VarBinViewArray>(&mut ctx).unwrap();
-
-    assert_eq!(
-        canonical.len(),
-        N,
-        "row count must survive canonicalization"
-    );
-    assert!(
-        canonical.data_buffers().len() >= 2,
-        "decoded heap exceeding MAX_BUFFER_LEN must roll over into multiple buffers, got {}",
-        canonical.data_buffers().len()
-    );
-    for (i, b) in canonical.data_buffers().iter().enumerate() {
-        assert!(
-            b.as_host().len() <= MAX_BUFFER_LEN,
-            "buffer {i} of {} bytes exceeds MAX_BUFFER_LEN",
-            b.as_host().len()
-        );
-    }
-
-    // Spot-check the endpoints and the rows straddling the rollover boundary, which is the first
-    // place the second buffer's offsets restart from zero.
-    let boundary = MAX_BUFFER_LEN / STRING_LEN;
-    for i in [0, boundary - 1, boundary, boundary + 1, N / 2, N - 1] {
-        assert_eq!(
-            canonical.bytes_at(i).as_slice(),
-            nth_string(i).as_slice(),
-            "value mismatch at row {i}"
-        );
-    }
 }
